feat: implement Order CRUD functionality with related services, controllers, and routes
- Added phase 6 documentation for Order CRUD implementation
- Created order DTOs, requests, and services for handling order operations
- Implemented order controller with handlers for creating, listing, retrieving, and canceling orders
- Registered order routes in the application
- Updated i18n files with new error messages for order-related operations
- Completed full CRUD implementation for orders, including transaction handling and status updates
- Added test execution report highlighting the need for unit and integration tests
- Documented completion status for all phases of the CRUD implementation plan

+++ backend/src/internal/models/user.go
@@ -6,8 +6,8 @@ import (
 	"gorm.io/gorm"
 )
 
-const USER_ROLE_ADMIN = 1
-const USER_ROLE_CUSTOMER = 2
+const USER_ROLE_ADMIN = int8(1)
+const USER_ROLE_CUSTOMER = int8(2)
 
 // AuthTokens chứa access token và refresh token
 type AuthTokens struct {
